// frontend/src/api/model/serviceOrderComposerSearchResult.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * KS Tech Service Management API
 * API for managing users, Products, and service orders in a service-center / repair business. Designed using API-first principles.
 * OpenAPI spec version: 1.0.0
 */
import type { ComposerContact } from './composerContact';
import type { ComposerProduct } from './composerProduct';
import type { ComposerUser } from './composerUser';
import type { ComposerUserProduct } from './composerUserProduct';

export interface ServiceOrderComposerSearchResult {
  user?: ComposerUser;
  contacts?: ComposerContact[];
  product?: ComposerProduct;
  userProduct?: ComposerUserProduct;
}